requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
discord==1.7.3
discord.py==1.7.3
dropbox==11.23.0
idna==3.3
multidict==5.2.0
Pillow==8.4.0
ply==3.11
pycparser==2.21
PyNaCl==1.4.0
PyYAML==6.0
requests==2.26.0
six==1.16.0
stone==3.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.2
